Address PR review feedback
- Revert recommended.mjs to original (cannot change recommended config)
- Keep noop processor unchanged; export template-lint-disable as a
  separate processor (ember/template-lint-disable) to avoid breaking
  existing gjs/gts users
- Wire hbs files to use the new processor by default in base configs
  (hbs had no prior config, so this is additive)
- Move hbs tests to a separate hbs-parser-test.js file
- Update gjs template-lint-disable tests to explicitly use the new
  processor via initESLintWithTemplateLintDisable()
