+++ lib/config-legacy/base.js
@@ -26,7 +26,7 @@ module.exports = {
     {
       files: ['**/*.hbs'],
       parser: 'ember-eslint-parser/hbs',
-      processor: 'ember/noop',
+      processor: 'ember/template-lint-disable',
     },
   ],
 };
