+++ lib/config/base.js
@@ -23,6 +23,6 @@ module.exports = [
     languageOptions: {
       parser: emberEslintParserHbs,
     },
-    processor: 'ember/noop',
+    processor: 'ember/template-lint-disable',
   },
 ];
