+++ lib/index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const requireIndex = require('requireindex');
+const noop = require('ember-eslint-parser/noop');
 const templateLintDisableProcessor = require('./processors/template-lint-disable');
 const pkg = require('../package.json'); // eslint-disable-line import/extensions
 
@@ -16,6 +17,7 @@ module.exports = {
   },
   processors: {
     // https://eslint.org/docs/developer-guide/working-with-plugins#file-extension-named-processor
-    noop: templateLintDisableProcessor,
+    noop,
+    'template-lint-disable': templateLintDisableProcessor,
   },
 };
