+++ lib/recommended.mjs
@@ -3,11 +3,9 @@ import baseRules from './recommended-rules.js';
 import gjsRules from './recommended-rules-gjs.js';
 import gtsRules from './recommended-rules-gts.js';
 import emberParser from 'ember-eslint-parser';
-import emberParserHbs from 'ember-eslint-parser/hbs';
 
 export const plugin = emberPlugin;
 export const parser = emberParser;
-export const hbsParser = emberParserHbs;
 
 export const base = {
   name: 'ember:base',
@@ -55,29 +53,14 @@ export const gts = {
   },
 };
 
-export const hbs = {
-  name: 'ember:hbs',
-  plugins: { ember: emberPlugin },
-  files: ['**/*.hbs'],
-  languageOptions: {
-    parser: emberParserHbs,
-  },
-  processor: 'ember/noop',
-  rules: {
-    ...base.rules,
-  },
-};
-
 export default {
   // Helpful utility exports
   plugin,
   parser,
-  hbsParser,
   // Recommended config sets
   configs: {
     base,
     gjs,
     gts,
-    hbs,
   },
 };
